+++ src/main/java/moze_intel/projecte/gameObjs/entity/PEProjectile.java
@@ -50,7 +50,7 @@ public abstract class PEProjectile extends EntityThrowable
 
     protected abstract void apply(RayTraceResult mop);
 
-    protected final boolean tryConsumeEmc(ItemPE consumeFrom, double amount)
+    protected final boolean tryConsumeEmc(ItemPE consumeFrom, long amount)
     {
         EntityPlayer player = ((EntityPlayer) getThrower());
         ItemStack found = PlayerHelper.findFirstItem(player, consumeFrom);
